refactor: move RollbackCounter to separate file
