+++ src/utils/slpParser.ts
@@ -15,8 +15,8 @@ import {
   PostFrameUpdateType,
   PreFrameUpdateType,
   RollbackFrames,
-  RollbackFramesType,
 } from "../types";
+import { RollbackCounter } from "./rollbackCounter";
 
 export const MAX_ROLLBACK_FRAMES = 7;
 
@@ -127,7 +127,11 @@ export class SlpParser extends EventEmitter {
   }
 
   public getRollbackFrames(): RollbackFrames {
-    return this.rollbackCounter.get();
+    return {
+      frames: this.rollbackCounter.getFrames(),
+      count: this.rollbackCounter.getCount(),
+      lengths: this.rollbackCounter.getLengths(),
+    };
   }
 
   public getFrame(num: number): FrameEntryType | null {
@@ -280,44 +284,5 @@ export class SlpParser extends EventEmitter {
       this.settingsComplete = true;
       this.emit(SlpParserEvent.SETTINGS, this.settings);
     }
-  }
-}
-
-class RollbackCounter {
-  private rollbackFrames: RollbackFramesType = {};
-  private rollbackFrameCount = 0;
-  private rollbackPlayerIdx: number | null = null; // for keeping track of rollbacks by following a single player
-  private lastFrameWasRollback = false;
-  private currentRollbackLength = 0;
-  private rollbackLengths: number[] = [];
-
-  public checkIfRollbackFrame(currentFrame: FrameEntryType | undefined, playerIdx: number) {
-    if (this.rollbackPlayerIdx === null) {
-      // we only want to follow a single player to avoid double counting. So we use whoever is on first.
-      this.rollbackPlayerIdx = playerIdx;
-    } else if (this.rollbackPlayerIdx !== playerIdx) {
-      return;
-    }
-
-    if (currentFrame) {
-      // frame already exists for currentFrameNumber so we must be rolling back
-      if (this.rollbackFrames[currentFrame.frame]) {
-        this.rollbackFrames[currentFrame.frame].push(currentFrame);
-      } else {
-        this.rollbackFrames[currentFrame.frame] = [currentFrame];
-      }
-      this.rollbackFrameCount++;
-      this.currentRollbackLength++;
-      this.lastFrameWasRollback = true;
-    } else if (this.lastFrameWasRollback) {
-      this.rollbackLengths.push(this.currentRollbackLength);
-      this.currentRollbackLength = 0;
-      this.lastFrameWasRollback = false;
-    }
-    return this.lastFrameWasRollback;
-  }
-
-  public get(): RollbackFrames {
-    return { frames: this.rollbackFrames, count: this.rollbackFrameCount, lengths: this.rollbackLengths };
   }
 }
